Specify md5 fingerprint_hash option in ssh-keygen (#11)
OpenSSH version 6.8 introduces, as reported in http://www.openssh.com/txt/release-6.8, the FingerprintHash option and the equivalent command-line flag to control the algorithm used for key fingerprints. With that change the default changes from MD5 to SHA256 and the format from hex to base64. This commit add the required flag to report the fingerprint in MD5 hex format.