Check use new flag (#14)
* Adding the new options use-md5

With this change we allow to specify a new option called 'use-md5',
when it is specified the ssh-keygen command used to verify the key
fingerprint will use the '-E' flag (fingerprintHash) which will force
the format of the fingerprint to be MD5. Please note that the
fingerprintHash is supported by OpenSSH versions equal or greater than
6.8.

This option by default is set to false.